+++ packages/runtime/src/lib/parsing.ts
@@ -1,5 +1,6 @@
 import { parsing } from "@algo-builder/web";
 import * as base32 from "hi-base32";
+import JSONbig from "json-bigint";
 
 import { RUNTIME_ERRORS } from "../errors/errors-list";
 import { RuntimeError } from "../errors/runtime-errors";
@@ -274,4 +275,45 @@ export function bigEndianBytesToBigInt(bytes: Uint8Array | Buffer): bigint {
 		return 0n;
 	}
 	return BigInt(buffToHex(bytes));
+}
+
+/**
+ * Parse String hex to bytes(represented as Uint8array)
+ * @param str
+ */
+export function strHexToBytes(str: string): Uint8Array {
+	return new Uint8Array(Buffer.from(str.slice(2), "hex"));
+}
+/*
+ * Function taken from algosdk.utils
+ * ConcatArrays takes n number arrays and returns a joint Uint8Array
+ * @param arrs - An arbitrary number of n array-like number list arguments
+ * @returns [a,b]
+ */
+export function concatArrays(...arrs: ArrayLike<number>[]) {
+	const size = arrs.reduce((sum, arr) => sum + arr.length, 0);
+	const c = new Uint8Array(size);
+
+	let offset = 0;
+	for (let i = 0; i < arrs.length; i++) {
+		c.set(arrs[i], offset);
+		offset += arrs[i].length;
+	}
+
+	return c;
+}
+
+/**
+ * assert if given string is a valid JSON object
+ * @param jsonString
+ */
+ export function assertJSON(jsonString: string, line: number): void {
+	const strictBigJSON = JSONbig({ strict: true });
+	try {
+		strictBigJSON.parse(jsonString);
+	} catch (e) {
+		throw new RuntimeError(RUNTIME_ERRORS.TEAL.INVALID_JSON_PARSING, {
+			line: line
+		});
+	}
 }