+++ packages/web/src/types.ts
@@ -5,6 +5,7 @@ import { Account as AccountSDK, LogicSigAccount, Transaction } from "algosdk";
 import * as z from "zod";
 
 import { WalletMultisigMetadata, WalletTransaction } from "./algo-signer-types";
+import { WAIT_ROUNDS } from "./lib/constants";
 import type { ASADefSchema, ASADefsSchema } from "./types-input";
 
 export type AccountAddress = string;
@@ -118,6 +119,7 @@ export type ExecParams =
 export enum SignType {
 	SecretKey,
 	LogicSignature,
+	MultiSignature
 }
 
 export enum TransactionType {
@@ -157,9 +159,16 @@ interface SignWithLsig {
 	args?: Uint8Array[];
 }
 
+export interface SignWithMultisig {
+	sign: SignType.MultiSignature;
+	mparams: WalletMultisigMetadata;
+	fromAccount?: AccountSDK;
+	fromAccountAddr: AccountAddress;
+}
+
 export type Lsig = SignWithLsig;
 
-export type Sign = SignWithSk | SignWithLsig;
+export type Sign = SignWithSk | SignWithLsig | SignWithMultisig;
 
 export type BasicParams = Sign & {
 	payFlags: TxParams;
@@ -234,11 +243,11 @@ export type UpdateAppParam = BasicParams &
 export type AppCallsParam = BasicParams &
 	AppOptionalFlags & {
 		type:
-			| TransactionType.CallApp
-			| TransactionType.ClearApp
-			| TransactionType.CloseApp
-			| TransactionType.DeleteApp
-			| TransactionType.OptInToApp;
+		| TransactionType.CallApp
+		| TransactionType.ClearApp
+		| TransactionType.CloseApp
+		| TransactionType.DeleteApp
+		| TransactionType.OptInToApp;
 		appID: number;
 	};
 
@@ -433,3 +442,5 @@ export interface HttpNetworkConfig {
 	token: string | AlgodTokenHeader | CustomTokenHeader;
 	httpHeaders?: { [name: string]: string };
 }
+
+export { WAIT_ROUNDS };